+++ caff-backend/pom.xml
@@ -12,10 +12,9 @@
 	</parent>
 	
 	<groupId>com.example</groupId>
-	<artifactId>distance_learning</artifactId>
+	<artifactId>caff_backend</artifactId>
 	<version>v1</version>
-	<name>distance_learning</name>
-	<description>Grocery app</description>
+	<name>caff_backend</name>
 
 	<properties>
 		<java.version>1.8</java.version>
