+++ caff-backend/src/main/java/com/example/cb/controller/AdminController.java
@@ -1,0 +1,17 @@
+package com.example.cb.controller;
+
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@CrossOrigin("*")
+@RestController
+@RequestMapping("/admin")
+public class AdminController {
+
+    @GetMapping("/welcome")
+    public String welcome() {
+        return "Welcome admin";
+    }
+}
