+++ caff-backend/src/main/java/com/example/cb/model/RoleEnum.java
@@ -1,0 +1,6 @@
+package com.example.cb.model;
+
+public enum RoleEnum {
+    ROLE_USER,
+    ROLE_ADMIN
+}
