+++ caff-backend/src/main/java/com/example/cb/model/User.java
@@ -1,0 +1,86 @@
+package com.example.cb.model;
+
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.*;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Setter
+@Getter
+@Table(	name = "users",
+		uniqueConstraints = {
+				@UniqueConstraint(columnNames = "username"),
+				@UniqueConstraint(columnNames = "email")
+		})
+public class User {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private Long user_id;
+	private String username;
+	private String email;
+	private String password;
+	@ManyToMany(fetch = FetchType.EAGER)
+	@JoinTable(	name = "user_role",
+			joinColumns = @JoinColumn(name = "user_id"),
+			inverseJoinColumns = @JoinColumn(name = "role_id"))
+	private List<Role> roles = new ArrayList<Role>();
+
+	public User(User user) {
+		this.user_id = user.user_id;
+		this.username = user.username;
+		this.email = user.email;
+		this.password = user.password;
+		this.roles = user.roles;
+	}
+
+	public User(String username, String email, String password) {
+		this.username = username;
+		this.email = email;
+		this.password = password;
+	}
+
+	public User() {
+	}
+
+	public String getEmail() {
+		return email;
+	}
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
+
+	public Long getUser_id() {
+		return user_id;
+	}
+
+	public void setUser_id(Long user_id) {
+		this.user_id = user_id;
+	}
+
+	public void setUsername(String username) {
+		this.username = username;
+	}
+
+	public String getPassword() {
+		return password;
+	}
+
+	public String getUsername() {
+		return username;
+	}
+
+	public List<Role> getRoles() {
+		return roles;
+	}
+
+	public void setPassword(String password) {
+		this.password = password;
+	}
+
+    public void setRoles(List<Role> roles) { this.roles = roles; }
+}
