+++ caff-backend/src/main/java/com/example/cb/payload/LoginRequest.java
@@ -1,0 +1,32 @@
+package com.example.cb.payload;
+
+//this is what the user sends in the post request
+public class LoginRequest {
+
+    private String username;
+    private String password;
+
+    public LoginRequest(String username, String password) {
+        this.username = username;
+        this.password = password;
+    }
+
+    public LoginRequest() {
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
