+++ caff-backend/src/main/java/com/example/cb/payload/MessageResponse.java
@@ -1,0 +1,17 @@
+package com.example.cb.payload;
+
+public class MessageResponse {
+    private String message;
+
+    public MessageResponse(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
