+++ caff-backend/src/main/java/com/example/cb/repository/RoleRepository.java
@@ -1,0 +1,14 @@
+package com.example.cb.repository;
+
+import java.util.Optional;
+
+import com.example.cb.model.Role;
+import com.example.cb.model.RoleEnum;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface RoleRepository extends JpaRepository<Role, Long> {
+    Optional<Role> findByName(RoleEnum name);
+}
