+++ caff-backend/src/main/java/com/example/cb/repository/UserRepository.java
@@ -1,0 +1,17 @@
+package com.example.cb.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.example.cb.model.User;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long>{
+	
+	User findByUsername(String username);
+
+	Boolean existsByUsername(String username);
+
+	Boolean existsByEmail(String email);
+}
+
