+++ caff-backend/src/main/java/com/example/cb/service/MyUserDetailsService.java
@@ -1,0 +1,29 @@
+package com.example.cb.service;
+
+import com.example.cb.model.MyUserDetails;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import com.example.cb.model.User;
+import com.example.cb.repository.UserRepository;
+
+
+@Service
+public class MyUserDetailsService implements UserDetailsService {
+
+	@Autowired
+	private UserRepository userRepository;
+
+	@Override
+	public UserDetails loadUserByUsername(final String username) {
+		final User user = userRepository.findByUsername(username);
+		if (user == null) {
+			throw new UsernameNotFoundException(username);
+		}
+		return new MyUserDetails(user);
+	}
+
+}
