+++ caff-backend/src/main/java/com/example/cb/util/JwtUtil.java
@@ -1,0 +1,64 @@
+package com.example.cb.util;
+
+import com.example.cb.model.MyUserDetails;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.Authentication;
+import org.springframework.stereotype.Service;
+
+import java.util.Date;
+
+@Service
+public class JwtUtil {
+
+    @Value("${jwtSecret}")
+    private String jwtSecret;
+
+    @Value("${jwtExpirationMs}")
+    private int jwtExpirationMs;
+
+    private static final Logger logger = LoggerFactory.getLogger(JwtUtil.class);
+
+    //takes the token and returns the username
+    public String getUserNameFromJwtToken(String token) {
+        return Jwts.parser().setSigningKey(jwtSecret).parseClaimsJws(token).getBody().getSubject();
+    }
+
+    //generate a jwt to the user
+    public String generateJwtToken(Authentication authentication) {
+
+        //get authenticated user's data
+        MyUserDetails userPrincipal = (MyUserDetails) authentication.getPrincipal();
+
+        //create token with the username, add exp time and sign
+        return Jwts.builder()
+                .setSubject((userPrincipal.getUsername()))
+                .setIssuedAt(new Date())
+                .setExpiration(new Date((new Date()).getTime() + jwtExpirationMs))
+                .signWith(SignatureAlgorithm.HS512, jwtSecret)
+                .compact();
+    }
+
+    public boolean validateJwtToken(String authToken) {
+        try {
+            Jwts.parser().setSigningKey(jwtSecret).parseClaimsJws(authToken);
+            return true;
+        } catch (SignatureException e) {
+            logger.error("Invalid JWT signature: {}", e.getMessage());
+        } catch (MalformedJwtException e) {
+            logger.error("Invalid JWT token: {}", e.getMessage());
+        } catch (ExpiredJwtException e) {
+            logger.error("JWT token is expired: {}", e.getMessage());
+        } catch (UnsupportedJwtException e) {
+            logger.error("JWT token is unsupported: {}", e.getMessage());
+        } catch (IllegalArgumentException e) {
+            logger.error("JWT claims string is empty: {}", e.getMessage());
+        }
+
+        return false;
+    }
+}
