--- caff-backend/src/main/java/com/example/dl/controller/AdminController.java
@@ -1,17 +1,0 @@
-package com.example.dl.controller;
-
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@CrossOrigin("*")
-@RestController
-@RequestMapping("/admin")
-public class AdminController {
-
-    @GetMapping("/welcome")
-    public String welcome() {
-        return "Welcome admin";
-    }
-}
