--- caff-backend/src/main/java/com/example/dl/controller/AuthController.java
@@ -1,134 +1,0 @@
-package com.example.dl.controller;
-
-import com.example.dl.model.Role;
-import com.example.dl.model.RoleEnum;
-import com.example.dl.payload.LoginRequest;
-import com.example.dl.payload.LoginResponse;
-import com.example.dl.model.MyUserDetails;
-import com.example.dl.model.User;
-import com.example.dl.payload.MessageResponse;
-import com.example.dl.payload.SignupRequest;
-import com.example.dl.repository.RoleRepository;
-import com.example.dl.repository.UserRepository;
-import com.example.dl.service.UserService;
-import com.example.dl.util.JwtUtil;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-@CrossOrigin("*")
-@RestController
-@RequestMapping("/auth")
-public class AuthController {
-
-    @Autowired
-    private AuthenticationManager authenticationManager;
-    @Autowired
-    private UserDetailsService userDetailsService;
-    @Autowired
-    private JwtUtil jwtUtil;
-    @Autowired
-    private UserRepository userRepository;
-    @Autowired
-    private RoleRepository roleRepository;
-    @Autowired
-    private UserService userService;
-    @Autowired
-    private PasswordEncoder encoder;
-
-    @PostMapping("/signin")
-    public ResponseEntity<?> authenticateUser(@RequestBody LoginRequest loginRequest) throws Exception {
-        Authentication authentication;
-
-        //check username and password
-        try {
-            authentication = authenticationManager.authenticate(
-                    new UsernamePasswordAuthenticationToken(loginRequest.getUsername(), loginRequest.getPassword()));
-
-        } catch (BadCredentialsException e) {
-            throw new Exception("Incorrect username or password", e);
-        }
-
-        //if the credentials are correct, authenticate the user and generate a token for it
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-        String jwt = jwtUtil.generateJwtToken(authentication);
-
-        //cast the user to MyUserDetails and set the roles
-        MyUserDetails userDetails = (MyUserDetails) authentication.getPrincipal();
-        List<String> roles = userDetails.getAuthorities().stream()
-                .map(item -> item.getAuthority())
-                .collect(Collectors.toList());
-
-        return ResponseEntity.ok(new LoginResponse(jwt, userDetails.getUser().getUser_id(), userDetails.getUsername(), userDetails.getUser().getEmail(), roles));
-    }
-    @PostMapping("/signup")
-    public ResponseEntity<?> registerUser(@RequestBody SignupRequest signUpRequest) {
-        //check if username or email already exists
-        if (userRepository.existsByUsername(signUpRequest.getUsername())) {
-            return ResponseEntity
-                    .badRequest()
-                    .body(new MessageResponse("Error: Username is already taken!"));
-        }
-
-        if (userRepository.existsByEmail(signUpRequest.getEmail())) {
-            return ResponseEntity
-                    .badRequest()
-                    .body(new MessageResponse("Error: Email is already in use!"));
-        }
-
-        //create new user's account
-        User user = new User(signUpRequest.getUsername(),
-                signUpRequest.getEmail(),
-                encoder.encode(signUpRequest.getPassword()));
-
-        //set the roles
-        Set<String> strRoles = signUpRequest.getRole();
-        List<Role> roles = new ArrayList<Role>();
-
-        //if no role was given, the default is user
-        if (strRoles == null) {
-            Role userRole = roleRepository.findByName(RoleEnum.ROLE_USER)
-                    .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
-            roles.add(userRole);
-        } else {
-
-            strRoles.forEach(role -> {
-                //check if admin
-                if ("admin".equals(role)) {
-                    Role adminRole = roleRepository.findByName(RoleEnum.ROLE_ADMIN)
-                            .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
-                    roles.add(adminRole);
-                }
-                //check if user
-                if ("user".equals(role))  {
-                    Role userRole = roleRepository.findByName(RoleEnum.ROLE_USER)
-                            .orElseThrow(() -> new RuntimeException("Error: Role is not found."));
-                    roles.add(userRole);
-                //check if none of them
-                } else {
-                    throw new RuntimeException("Error: Wrong Role.");
-                }
-            });
-        }
-
-        //give the role to the user and save it
-        user.setRoles(roles);
-        userService.save(user);
-
-        return ResponseEntity.ok(new MessageResponse("User registered successfully!"));
-    }
-
-}
