--- caff-backend/src/main/java/com/example/dl/model/MyUserDetails.java
@@ -1,77 +1,0 @@
-package com.example.dl.model;
-
-
-import java.util.Collection;
-
-import java.util.Objects;
-import java.util.stream.Collectors;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import com.example.dl.model.User;
-
-import lombok.Getter;
-import lombok.Setter;
-
-
-@Getter
-@Setter
-public class MyUserDetails  implements UserDetails{
-
-	User user;
-	public MyUserDetails(User user) {
-		this.user = user;
-	}
-
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		return this.user.getRoles().stream()
-				.map(role -> new SimpleGrantedAuthority(role.getRole().name()))
-				.collect(Collectors.toList());
-
-	}
-
-	@Override
-	public String getPassword() {
-		return this.user.getPassword();
-	}
-
-	@Override
-	public String getUsername() {
-		return this.user.getUsername();
-	}
-
-	public User getUser(){ return this.user; }
-
-	@Override
-	public boolean isAccountNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isAccountNonLocked() {
-		return true;
-	}
-
-	@Override
-	public boolean isCredentialsNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isEnabled() {
-		return true;
-	}
-
-	@Override
-	public boolean equals(Object o) {
-		if (this == o)
-			return true;
-		if (o == null || getClass() != o.getClass())
-			return false;
-		MyUserDetails user = (MyUserDetails) o;
-		return Objects.equals(this.user.getUser_id(), user.user.getUser_id());
-	}
-}
