--- caff-backend/src/main/java/com/example/dl/model/RoleEnum.java
@@ -1,6 +1,0 @@
-package com.example.dl.model;
-
-public enum RoleEnum {
-    ROLE_USER,
-    ROLE_ADMIN
-}
