--- caff-backend/src/main/java/com/example/dl/payload/MessageResponse.java
@@ -1,17 +1,0 @@
-package com.example.dl.payload;
-
-public class MessageResponse {
-    private String message;
-
-    public MessageResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
