--- caff-backend/src/main/java/com/example/dl/repository/RoleRepository.java
@@ -1,14 +1,0 @@
-package com.example.dl.repository;
-
-import java.util.Optional;
-
-import com.example.dl.model.Role;
-import com.example.dl.model.RoleEnum;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
-    Optional<Role> findByName(RoleEnum name);
-}
