--- caff-backend/src/main/java/com/example/dl/repository/UserRepository.java
@@ -1,17 +1,0 @@
-package com.example.dl.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.example.dl.model.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long>{
-	
-	User findByUsername(String username);
-
-	Boolean existsByUsername(String username);
-
-	Boolean existsByEmail(String email);
-}
-
