--- caff-backend/src/main/java/com/example/dl/security/SecurityConfig.java
@@ -1,67 +1,0 @@
-package com.example.dl.security;
-
-import com.example.dl.service.MyUserDetailsService;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import javax.servlet.Filter;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
-
-	@Autowired
-	private MyUserDetailsService userDetailsService;
-	@Autowired
-	private Filter jwtRequestFilter;
-	@Autowired
-	AuthEntryPointJwt unauthorizedHandler;
-
-	@Override
-	protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-		auth.userDetailsService(userDetailsService).passwordEncoder(passwordEncoder());
-	}
-
-	@Override
-	@Bean
-	public AuthenticationManager authenticationManagerBean() throws Exception {
-		return super.authenticationManagerBean();
-	}
-
-	@Bean
-	public BCryptPasswordEncoder passwordEncoder() {
-		return new BCryptPasswordEncoder();
-	}
-
-	@Override
-	protected void configure(HttpSecurity http) throws Exception {
-		http.cors().and().csrf().disable()
-				//will be triggered anytime unauthenticated user requests a secured HTTP resource
-				.exceptionHandling().authenticationEntryPoint(unauthorizedHandler).and()
-				//tells spring security to not create session automatically
-				.sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS).and()
-				.authorizeRequests()
-				//sign in and sign up is available for everyone
-				.antMatchers("/auth/**").permitAll()
-				//only admin
-				.antMatchers("/admin/**").hasRole("ADMIN")
-				//only authenticated users
-				.anyRequest().authenticated()
-				.and();
-			    //call this filter before the auth method, does the session (if the token is valid)
-				http.addFilterBefore(jwtRequestFilter, UsernamePasswordAuthenticationFilter.class);
-	}
-}
