+++ caff-backend/src/test/java/com/example/caffbackend/CaffBackendApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.caffbackend;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CaffBackendApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
