--- caff-backend/src/test/java/com/example/distancelearning/DistanceLearningApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.distancelearning;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DistanceLearningApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
